minor fixes in AddBookDTo, UpdateBookDTO

+++ Backend/Backend/DTOs/Admin/Book/UpdateBookDTO.cs
@@ -11,9 +11,9 @@
         public string? Language { get; set; }
         public string? Format { get; set; }
         public string? Description { get; set; }
-        public float? Price { get; set; }
+        public decimal? Price { get; set; }
         public int? Stock { get; set; }
-        public float? Discount { get; set; }
+        public decimal? Discount { get; set; }
 
     }
 }
